ACT4 - Amador-S - Refactor code to separate component user
