+++ vite-project/src/components/User.css
@@ -1,0 +1,11 @@
+.avatar {
+  display: flex;
+  justify-content: center;
+}
+.name {
+  color: black;
+}
+.card-container {
+  display: flex;
+  justify-content: center;
+}
