+++ vite-project/src/components/UserAvatar.tsx
@@ -1,0 +1,12 @@
+import Avatar from '@mui/material/Avatar';
+
+function UserAvatar() {
+    return (
+        <>
+            <div className="avatar">
+                <Avatar sx={{ width: 150, height: 150, fontSize: '50px' }}>AS</Avatar>
+            </div>
+        </>
+    )
+}
+export default UserAvatar
